MAJOR UPDATE: new system for error modelling and Survey.write2protocol()
Survey.basicFilter():
--------------------
The default behavior is now to NOT discard the non-paired measurements by default but keep them all. If the dataset contains some reciprocal then an error model can be fitted on them and applied to all other measurements (even non-paired).
Survey.basicFilter() now removes NaN, Inf, duplicates, strange quadrupoles where current electrode are also potential electrode and dummy (with the rule M < N for 2D only) but NOT non-paired quadrupoles.

Survey.write2protocol():
------------------------
Concerning the error model. The old argument errTyp and errTypIP of Survey.write2protocol() are removed and replace by a simple err=True/False argument. The default is that there is no error model (err=False). If err=True then the code will look into the phaseError and resError columns of Survey.df, if they contains non-NaN it will use them if not, it will throw an error.

Because of those two updates and the new R2.matchSurveys() the `trim` and `keepAll` arguments of R2.createSurvey(), R2.createTimeLapseSurvey() and R2.createBatchSurvey() are removed.
(as the default is now to no trim until we know what reg_mode is used and to keep all even non-paired measurements).