Mesh background region optimisations. On Andy's advice the extents of the mesh are defined by the largest dipole length.
The arguments is called ```dp_len```, changing this will affect the lateral and depth extents of the mesh.

For 2 (and 3) D the lateral flanks are defined as 5*dp_len. Depth of the mesh is defined as
doi - (3*dp_len).

By default the dp_len is assumed to be the largest possible horizontal distance between 2 electrodes.